moving exception to test
